requirements.txt
matplotlib
tqdm
tensorboardX
torchsummary
Pillow
pytorch-lightning==1.5.10
timm
randaugment
albumentations
git+https://github.com/adam-mehdi/MuarAugment.git